Refactor the utilities file into utilities folder
