+++ transformer_lens/utilities/components_utils.py
@@ -1,0 +1,108 @@
+"""activation_utils.
+
+This module contains utility functions related to model components
+"""
+
+from __future__ import annotations
+
+import re
+from typing import Optional, Union
+
+
+def get_act_name(
+    name: str,
+    layer: Optional[Union[int, str]] = None,
+    layer_type: Optional[str] = None,
+):
+    """
+    Helper function to convert shorthand to an activation name. Pretty hacky, intended to be useful for short feedback
+    loop hacking stuff together, more so than writing good, readable code. But it is deterministic!
+
+    Returns a name corresponding to an activation point in a TransformerLens model.
+
+    Args:
+         name (str): Takes in the name of the activation. This can be used to specify any activation name by itself.
+         The code assumes the first sequence of digits passed to it (if any) is the layer number, and anything after
+         that is the layer type.
+
+         Given only a word and number, it leaves layer_type as is.
+         Given only a word, it leaves layer and layer_type as is.
+
+         Examples:
+             get_act_name('embed') = get_act_name('embed', None, None)
+             get_act_name('k6') = get_act_name('k', 6, None)
+             get_act_name('scale4ln1') = get_act_name('scale', 4, 'ln1')
+
+         layer (int, optional): Takes in the layer number. Used for activations that appear in every block.
+
+         layer_type (string, optional): Used to distinguish between activations that appear multiple times in one block.
+
+    Full Examples:
+
+    get_act_name('k', 6, 'a')=='blocks.6.attn.hook_k'
+    get_act_name('pre', 2)=='blocks.2.mlp.hook_pre'
+    get_act_name('embed')=='hook_embed'
+    get_act_name('normalized', 27, 'ln2')=='blocks.27.ln2.hook_normalized'
+    get_act_name('k6')=='blocks.6.attn.hook_k'
+    get_act_name('scale4ln1')=='blocks.4.ln1.hook_scale'
+    get_act_name('pre5')=='blocks.5.mlp.hook_pre'
+    """
+    if ("." in name or name.startswith("hook_")) and layer is None and layer_type is None:
+        # If this was called on a full name, just return it
+        return name
+    match = re.match(r"([a-z]+)(\d+)([a-z]?.*)", name)
+    if match is not None:
+        name, layer, layer_type = match.groups(0)  # type: ignore
+
+    layer_type_alias = {
+        "a": "attn",
+        "m": "mlp",
+        "b": "",
+        "block": "",
+        "blocks": "",
+        "attention": "attn",
+    }
+
+    act_name_alias = {
+        "attn": "pattern",
+        "attn_logits": "attn_scores",
+        "key": "k",
+        "query": "q",
+        "value": "v",
+        "mlp_pre": "pre",
+        "mlp_mid": "mid",
+        "mlp_post": "post",
+    }
+
+    layer_norm_names = ["scale", "normalized"]
+
+    if name in act_name_alias:
+        name = act_name_alias[name]
+
+    full_act_name = ""
+    if layer is not None:
+        full_act_name += f"blocks.{layer}."
+    if name in [
+        "k",
+        "v",
+        "q",
+        "z",
+        "rot_k",
+        "rot_q",
+        "result",
+        "pattern",
+        "attn_scores",
+    ]:
+        layer_type = "attn"
+    elif name in ["pre", "post", "mid", "pre_linear"]:
+        layer_type = "mlp"
+    elif layer_type in layer_type_alias:
+        layer_type = layer_type_alias[layer_type]
+
+    if layer_type:
+        full_act_name += f"{layer_type}."
+    full_act_name += f"hook_{name}"
+
+    if name in layer_norm_names and layer is None:
+        full_act_name = f"ln_final.{full_act_name}"
+    return full_act_name
